pyproject.toml
[tool.poetry]
name = "hexlet-code"
version = "0.1.0"
description = ""
authors = ["ilnarkz <ilnar-16_04@mail.ru>"]

[tool.poetry.dependencies]
python = "^3.8"
django-bootstrap4 = "^22.2"

[tool.poetry.dev-dependencies]
flake8 = "^5.0.4"
pytest = "^7.1.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
